requirements.txt
Flask
waitress
google-cloud-firestore
google-cloud-storage
google-cloud-logging
google-cloud-error-reporting
google-cloud-pubsub
six
pillow
